Add a new API function (AssertEx) and fix an off-by-one bug
The new API function replaces the original Assert function, whose
return values are user-definable but default to PAWNTEST_CASE_FAILED
and PAWNTEST_CASE_PASSED, depending on the evaulation. AssertEx is
now a wrapper function around Assert with false and true as its
default values. This change is reflected in test-suite-test.pwn
as well.

Fix an off-by-one bug. This bug was discovered by test-suite-test.pwn.